+++ .vscode/settings.json
@@ -1,16 +1,16 @@
 {
     "workbench.colorCustomizations": {
-        "editorGroup.border": "#7f40bf",
-        "panel.border": "#7f40bf",
-        "sideBar.border": "#7f40bf",
-        "statusBar.background": "#663399",
-        "statusBar.foreground": "#e7e7e7",
-        "statusBarItem.hoverBackground": "#7f40bf",
-        "tab.activeBorder": "#7f40bf",
-        "titleBar.activeBackground": "#663399",
-        "titleBar.activeForeground": "#e7e7e7",
-        "titleBar.inactiveBackground": "#66339999",
-        "titleBar.inactiveForeground": "#e7e7e799"
+        "editorGroup.border": "#88d7ea",
+        "panel.border": "#88d7ea",
+        "sideBar.border": "#88d7ea",
+        "statusBar.background": "#5dc9e2",
+        "statusBar.foreground": "#15202b",
+        "statusBarItem.hoverBackground": "#32bbda",
+        "tab.activeBorder": "#88d7ea",
+        "titleBar.activeBackground": "#5dc9e2",
+        "titleBar.activeForeground": "#15202b",
+        "titleBar.inactiveBackground": "#5dc9e299",
+        "titleBar.inactiveForeground": "#15202b99"
     },
-    "peacock.color": "#639"
+    "peacock.color": "#5dc9e2"
 }
