+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/CharpEngine.csproj",
+                "${workspaceFolder}/src/CharpEngine.Engine/CharpEngine.Engine.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/CharpEngine.csproj",
+                "${workspaceFolder}/src/CharpEngine.Engine/CharpEngine.Engine.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,7 +32,7 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/CharpEngine.csproj",
+                "${workspaceFolder}/src/CharpEngine.Engine/CharpEngine.Engine.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
