+++ CharpEngine.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30503.244
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CharpEngine", "CharpEngine.csproj", "{AFCBC255-DE48-433D-9C01-0FC125D41F52}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CharpEngine", "src/CharpEngine.Engine/CharpEngine.Engine.csproj", "{AFCBC255-DE48-433D-9C01-0FC125D41F52}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
