--- Chunk.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CharpEngine
-{
-    public class Chunk
-    {
-
-    }
-}
