--- Voxel.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Text;
-using OpenTK.Graphics.OpenGL;
-using OpenTK.Mathematics;
-
-namespace CharpEngine
-{
-    public class Voxel
-    {
-        public bool IsActive { get; set; }
-        public Color Color { get; set; }
-
-        public Voxel()
-        {
-
-        }
-    }
-}
