+++ src/CharpEngine.Engine/ICamera.cs
@@ -1,0 +1,10 @@
+using OpenTK.Mathematics;
+
+namespace CharpEngine
+{
+    public interface ICamera
+    {
+         public Matrix4 GetViewMatrix();
+         public Matrix4 GetProjectionMatrix();
+    }
+}
