+++ src/CharpEngine.Engine/Voxel.cs
@@ -1,0 +1,12 @@
+﻿
+namespace CharpEngine
+{
+    public class Voxel
+    {
+        public Voxel()
+        {
+        }
+        public bool IsActive { get; set; }
+        public VoxelType Type { get; set; }
+    }
+}
