+++ src/CharpEngine.Engine/VoxelType.cs
@@ -1,0 +1,8 @@
+namespace CharpEngine
+{
+    public enum VoxelType: byte
+    {
+        Sand,
+        Dirt
+    }
+}
